pyproject.toml
[project]
name = "python-edition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-cloud-bigquery>=3.38.0",
    "load-dotenv>=0.1.0",
]
